requirements.txt
appdirs==1.4.4
attrs==23.2.0
Automat==22.10.0
azure-core==1.30.1
azure-identity==1.15.0
azure-storage-blob==12.19.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
DateTime==5.5
dnspython==2.6.1
filelock==3.13.1
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.50.0
frozendict==2.4.0
fsspec==2024.3.1
gpxpy==1.6.2
h11==0.14.0
html5lib==1.1
huggingface-hub==0.22.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
multitasking==0.0.11
networkx==3.2.1
numpy==1.26.4
onnxruntime==1.17.1
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pandas_ta==0.3.14b0
parsel==1.9.0
peewee==3.17.1
pillow==10.3.0
plotly==5.20.0
portalocker==2.8.2
Protego==0.3.0
protobuf==5.26.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyJWT==2.8.0
pymongo==4.6.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
queuelib==1.6.2
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
Scrapy==2.11.1
scrapy-selenium==0.0.7
seaborn==0.13.2
selenium==4.18.1
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tldextract==5.1.2
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.2
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
w3lib==2.1.2
webencodings==0.5.1
Werkzeug==3.0.1
wsproto==1.2.0
yfinance==0.2.37
zope.interface==6.2
